backend/requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document Processing - Pinned versions to avoid conflicts
sentence-transformers==2.2.2
huggingface-hub==0.16.4  # Specific version that works with sentence-transformers
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.3

# Vector Database
chromadb==0.4.15

# Text Processing
langchain==0.0.350

# AI Models
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
safetensors==0.4.1
tokenizers==0.15.1
google-generativeai>=0.3.0

# Additional dependencies
numpy==1.24.3
scikit-learn==1.3.2
tqdm==4.66.1
